refactor: popover moved into its own hook file
